// projects/apppensionista/src/app/features/autorizar-datos/models/informacion.model.ts
export class InformacionModel{
    codigo: number = 0;
    apePaterno: string = "";
    apeMaterno: string = "";
    nomPensionista: string = "";
    sigTipoDocumento: string = "";
    flgEsPensionista: string = "";
    idTipoDocuIdentidad: string = "";
    numDocuIdentidad: string = "";
    desTipoDocumento: string = "";
    desCuenta: string = "";
    prestacion: string = "";
    TipoDerecho: number = 0;
    desRegimen: string = "";
    desEntidad: string = "";
    fechaIniPension: string = "";
    estadoCuenta: string ="";
    indProvisional: string ="";
    msjPensionista: string = "";
    sexo: string ="";
}