// projects/apppensionista/src/app/features/autorizar-datos/models/registrar.model.ts
export class RegistrarModel{
    esAutorizacion: boolean = false;
    tipoDocumento: string | undefined;
    numeroDocumento: string | undefined;
    regimen: string = "";
    numeroRuc: string = "";
    codigoCooperativa: string = "";
    descripcionCooperativa: string = "";
    tipoAutorizacion: string = "BL";
    usuarioCreacion: string | undefined;
    ipRegistro: string = "";
    cuentaPension: string = "";
    prestacion: string = "";
    tipoCooperativa: string ="";
}